changed #bins spe_fit parameter from float to int, enabled gain cal, corrected spe_fit bin count

+++ private/dom-cal/spefit.h
@@ -3,4 +3,4 @@
 #define SPE_FIT_MAX_ITER         20
 #define SPE_FIT_CHISQ_DONE     0.01
 
-int spe_fit (float *xdata, float *ydata, float pts, float *fit_params);
+int spe_fit(float *xdata, float *ydata, int pts, float *fit_params);
